+++ src/com/gregswebserver/ld28/game/level/tile/PathTile.java
@@ -1,10 +1,11 @@
 package com.gregswebserver.ld28.game.level.tile;
 
 import com.gregswebserver.ld28.graphics.sprite.Sprite;
+import com.gregswebserver.ld28.util.vectors.Vector2d;
 import com.gregswebserver.ld28.util.vectors.Vector2i;
 
 public class PathTile extends Tile {
-    public PathTile(Vector2i position) {
+    public PathTile(Vector2d position) {
         super(position);
     }
 }
