+++ src/com/gregswebserver/ld28/game/level/tile/WallTile.java
@@ -1,9 +1,9 @@
 package com.gregswebserver.ld28.game.level.tile;
 
-import com.gregswebserver.ld28.util.vectors.Vector2i;
+import com.gregswebserver.ld28.util.vectors.Vector2d;
 
 public class WallTile extends Tile {
-    public WallTile(Vector2i position) {
+    public WallTile(Vector2d position) {
         super(position);
     }
